Moved model components to separate python modules
