+++ src/llama_jax/__init__.py
@@ -1,5 +1,6 @@
-from . import model
+from . import model, normalization
 
 __all__ = [
     "model",
+    "normalization",
 ]
