+++ src/llama_jax/head.py
@@ -1,0 +1,19 @@
+"""Head."""
+
+from typing import NamedTuple
+
+from jax import Array
+
+from .normalization import RMSNorm
+
+__all__ = [
+    "Head",
+]
+
+
+class Head(NamedTuple):
+    """Head state."""
+
+    norm: RMSNorm
+
+    output: Array
